avip/requirements.txt
Pillow
numpy
matplotlib
tqdm
fontTools
prettytable